# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Provides data from the Minecart Rapid Transit (MRT) server.",
        "title": "Minecart Rapid Transit Server API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "https://api.minecartrapidtransit.net",
    "basePath": "/api/v1",
    "paths": {
        "/companies": {
            "get": {
                "description": "List all companies (defined in https://github.com/Frumple/mrt-api/blob/main/data/companies.yml).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "List all companies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Company"
                            }
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "description": "Get company by ID (defined in https://github.com/Frumple/mrt-api/blob/main/data/companies.yml).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Companies"
                ],
                "summary": "Get company by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Company"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/warps": {
            "get": {
                "description": "List all warps.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Warps"
                ],
                "summary": "List all warps",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by warp name.",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by player UUID (with or without hyphens).",
                        "name": "player",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by company ID (from /companies).",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by world ID (from /worlds).",
                        "name": "world",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by 'name', 'creation_date', or 'visits'.",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by 'asc' (ascending) or 'desc' (descending).",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of warps returned.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of warps to skip before returning. Can only be used if the limit query parameter is set.",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Warp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/warps/{id}": {
            "get": {
                "description": "Get warp by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Warps"
                ],
                "summary": "Get warp by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warp ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Warp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        },
        "/worlds": {
            "get": {
                "description": "List all worlds (defined in https://github.com/Frumple/mrt-api/blob/main/data/worlds.yml).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worlds"
                ],
                "summary": "List all worlds",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.World"
                            }
                        }
                    }
                }
            }
        },
        "/worlds/{id}": {
            "get": {
                "description": "Get world by ID (defined in https://github.com/Frumple/mrt-api/blob/main/data/worlds.yml).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Worlds"
                ],
                "summary": "Get world by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "World ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.World"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Company": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pattern": {
                    "type": "string"
                }
            }
        },
        "main.Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "main.Warp": {
            "type": "object",
            "properties": {
                "creationDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pitch": {
                    "type": "number"
                },
                "playerUUID": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "visits": {
                    "type": "integer"
                },
                "welcomeMessage": {
                    "type": "string"
                },
                "worldUUID": {
                    "type": "string"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "yaw": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "main.World": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    },
    "externalDocs": {
        "description": "GitHub Repository",
        "url": "https://github.com/Frumple/mrt-api"
    }
}